add controller and service for Rules

+++ core/src/main/java/com/lromal/rulesStatistic/repository/RuleRepository.java
@@ -4,6 +4,6 @@ import com.lromal.rulesStatistic.model.Rule;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface RuleRepository extends CrudRepository<Rule, Integer> {
+public interface RuleRepository extends CrudRepository<Rule, Long> {
 
 }
